Separação de gráficos em components, organização da página Dashboard.
